pep8 rename private modules
