+++ src/cosmology/api/_core.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Protocol, runtime_checkable
 from cosmology.api._array_api import ArrayT_co
 
 if TYPE_CHECKING:
-    from .namespace import CosmologyAPINamespace
+    from ._namespace import CosmologyAPINamespace
 
 __all__: list[str] = []
