+++ src/cosmology/api/_namespace.py
@@ -15,7 +15,7 @@ from typing import TYPE_CHECKING, Protocol, runtime_checkable
 __all__: list[str] = []
 
 if TYPE_CHECKING:
-    from cosmology.api.constants import CosmologyConstantsAPINamespace
+    from cosmology.api._constants import CosmologyConstantsAPINamespace
 
 
 @runtime_checkable
