+++ src/cosmology/api/compat/__init__.py
@@ -1,7 +1,7 @@
 """The Cosmology API standard for compatability wrappers."""
 
-from cosmology.api.compat.core import CosmologyWrapperAPI
-from cosmology.api.compat.standard import StandardCosmologyWrapperAPI
+from cosmology.api.compat._core import CosmologyWrapperAPI
+from cosmology.api.compat._standard import StandardCosmologyWrapperAPI
 
 __all__ = [
     "CosmologyWrapperAPI",
