+++ src/cosmology/api/compat/_standard.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 from typing import Protocol
 
 from cosmology.api._array_api import ArrayT
-from cosmology.api.compat.core import CosmologyWrapperAPI
-from cosmology.api.standard import StandardCosmologyAPI
+from cosmology.api._standard import StandardCosmologyAPI
+from cosmology.api.compat._core import CosmologyWrapperAPI
 
 __all__: list[str] = []
